Move produceNextState from backtest to dedicated file
refactor
